Move measurements into a new package
This shouldn't depend upon JavaFX, so in the future can be moved up to a more core module.
