+++ qupath-gui-fx/src/main/java/qupath/lib/gui/measure/measurements/PixelClassifierMeasurementBuilder.java
@@ -1,4 +1,4 @@
-package qupath.lib.gui.measure;
+package qupath.lib.gui.measure.measurements;
 
 import qupath.lib.objects.PathObject;
 import qupath.opencv.ml.pixel.PixelClassificationMeasurementManager;
