+++ qupath-gui-fx/src/main/java/qupath/lib/gui/measure/measurements/StringMeasurementBuilder.java
@@ -1,4 +1,4 @@
-package qupath.lib.gui.measure;
+package qupath.lib.gui.measure.measurements;
 
 public interface StringMeasurementBuilder extends MeasurementBuilder<String> {
